requirements.txt
requests
aiohttp
enum34
typing
python-telegram-bot
redis